+++ apps/scoutgame/components/common/ClientGlobals.tsx
@@ -5,8 +5,8 @@ import { usePageView } from '@connect-shared/hooks/usePageView';
 import { useDatadogLogger } from '@root/hooks/useDatadogLogger';
 
 // instantiate global hooks for the client-side only
-export function ClientGlobal({ userId }: { userId?: string }) {
-  useDatadogLogger({ service: 'sunnyawards-browser', userId });
+export function ClientGlobals({ userId }: { userId?: string }) {
+  useDatadogLogger({ service: 'scoutgame-browser', userId });
   usePageView();
   useDarkTheme();
   return null;
