+++ apps/scoutgame/components/common/WalletLogin/WagmiProvider.tsx
@@ -1,11 +1,12 @@
 'use client';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { headers } from 'next/headers';
 import { useState } from 'react';
 import type { State } from 'wagmi';
-import { WagmiProvider as OriginalWagmiProvider } from 'wagmi';
+import { cookieToInitialState, WagmiProvider as OriginalWagmiProvider } from 'wagmi';
 
-import { getConfig } from '../../lib/auth/wagmiConfig';
+import { getConfig } from 'lib/auth/wagmiConfig';
 
 export function WagmiProvider({ children, initialState }: { children: React.ReactNode; initialState?: State }) {
   const [config] = useState(() => getConfig());
