+++ apps/scoutgame/components/common/WalletLogin/WalletProvider.tsx
@@ -1,0 +1,12 @@
+import { headers } from 'next/headers';
+import { cookieToInitialState } from 'wagmi';
+
+import { getConfig } from 'lib/auth/wagmiConfig';
+
+import { WagmiProvider } from './WagmiProvider';
+
+export function WalletProvider({ children }: { children: React.ReactNode }) {
+  const wagmiInitialState = cookieToInitialState(getConfig(), headers().get('cookie'));
+
+  return <WagmiProvider initialState={wagmiInitialState}>{children}</WagmiProvider>;
+}
