+++ apps/scoutgame/components/home/ComingSoon.tsx
@@ -1,0 +1,7 @@
+export function ComingSoon() {
+  return (
+    <div>
+      <h1>Coming Soon!</h1>
+    </div>
+  );
+}
